Fixes the DampedNewton logic to take the maxstep, not one more
While working on device masking in the nonlinear solver it was
observed that the internal Newton solver (DampedNewton) will take
one more step than the maximum number of steps.  This logic is being
corrected in this commit, which is simple.  However, than change
resulted in a couple convergence failures and comparison failures
in the regression tests that will also be fixed.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/329