Change formatting of column headers for .PRINT SENS
The notation for the sensitivity analysis column headers is different
between AC and DC in the .PRINT SENS output. This change is intended
to make the formatting in the sensitivity output more consistent across
the different analysis types.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/270

Change case formatting of headers for AC SENS

The command parser that builds the sensitivity Option Block (extracSENSData)
is not formatting the parameters/values from the parsed netlist line before
adding it to the option block. Other command parsers convert netlist text to
all upper case before assigning it to an option block parameter. There is
some case formatting logic for sensitivity parameters further downstream when
the option block is processed, but it doesn’t cover all cases. This change
adds upper case formatting to the sensitivity command parser (and removes the
redundant downstream formatting).

Start consolidating objvars and acobjfunc handling

I noted that the ac sensitivity outputter is hardcoding voltage
component operators (VR, VI, VP, VM) around the objective function
name in the column headers of the output files. This works for
objectives specified using objvars, but objectives specified using
acobjfunc may not be voltages. The suggested approach is to use newer
component operators that do not imply the value type of the objective
function (i.e. Re(), Im(), Mag(), Ph()). This will correct the format
for acobjfunc, but for objvars the value type still needs to be added.
Additional logic is added to the ac sensitivity command processor to
combine objective function definitions from both acobjfunc and objvars
(instead of having seperate logic paths).

Remove objvars specific code from .sens analysis

The initial AC sensitivity routines were implemented before the
expression library supported complex numbers. For this solution,
only objvars was supported in .SENS lines for specifying objective
functions. After the expression library was updated, a new option was
added in the AC sensitivity analysis for specifying objective functions
as expressions (acobjfunc). Previous commits for this issue updated the
parsing logic for objvars parameters so that they are handled the same
as acobjfunc parameters. The changes in this commit remove a lot of the
initial objvars logic that is no longer executed.

Release note and RG updates for .sens output changes

Added release note for Gitlab issue 270. Also updated the Print
Sensitivity tables in the Reference Guide. A section of the table
describes the columns for the sensitivity data in the output file. This
should make the notation in the table a closer match to the actual
output. Noted that AC sensitivities table does not mention that the
complex and polar forms included in the output