Python 3 support and access new C function
This commit adds Python 3 support.  In the xyce_interface.py, many
byte-arrays that held c-strings were conveted to python strings using
the default encoding.  This was ok in Python 2.7 but in Python 3.x
the resulting string will keep it's leading "b" character if the
encoding wasn't set.  Thus for Python 3 support, numerous
decode('utf-8') calls were added when c-strings are conveted to Python
strings.

This commit also exposes the XyceCInterface function
xyce_getTimeVoltagePairsADCLimitData which allows the Python interface
to return a properly sized time-voltage pair arrays for the ADC's in
a system without assuming each ADC has the same number of time-voltage
pairs.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/351
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/348