fix string to bytearray encodeing in python
In the python to c interface layer there were a few more points
where a python string was being converted to a c-string and
needed the encoding specified to work with Python 3.

Also, the routine xyce_getDeviceParamVal() should zero out
the value of the double passed into it if a the requrested
device parameter is not found.  It was not doing this and tests
for this were passing on linux but not my mac.  On linux
the returned, but uninitialized value was zero.  Making this
function set the value when the device paramter is not found allows
the test to work correctly on linux and mac's.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/351