Add memory safe ADC state var access.
In the same way that xyce_getTimeVoltagePairsADC() in XyceCInterface
could overwrite the memory areas passed into it, the
xyce_getTimeStatePairsADC() function had the same problem.  This
commit adds in a LimitedData version of the StatePairsADC() function
so that hard limits on buffer sizes can be respected. The
function is also available to python through the xyce_interface.py
file.  Within xyce_interface.py, the limits on time points and ADC
name length that were hard coded into the python routines are now
settable as function parameters.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/348