Improve to expression breakpoint robustness
Issue #340
"Bug in breakpointing, when an expression combines a ternary and a
table"

I worked on this issue a few weeks ago, but my work wasn't quite
sufficient to close it.  This commit should do it.

The "computeBreakPoint" function was not robust under very unusual
circumstances.  One circumstance was if the condition for the breakpoint
was already satisfied.  Prior to this commit, this could result in a new
breakpoint being created for the current time point.  This didn't really
make any sense, b/c if you've already landed on the breakpoint, you don't
need to add it again to the breakpoint container.

If this happens, then the best case scenario is that it will get
purged later in the step error control class, and the worst case is
that the time integrator will get stuck.

This commit fixes this (rare) robustness problem.