Make VBIC "OFF" simply skip junction voltage initialization
When Xyce/Xyce#354 was opened, I interpreted it as a request to add an
OFF parameter that did basically what that parameter does in the level
1 SPICE-Gummel-Poon (SGP) BJT model.

SGP does something funky with OFF:  it zeros all the
voltage drops (base'-emitter', base'-collector', and base-collector')
when "initFix" is active (MODEINITFIX in SPICE, initFixFlag in Xyce).
That is active during the DCOP whenever any device in the circuit is
not converged (including the SGP device itself).

When I implemented *that* in VBIC it behaved badly on the simplest
circuit I could find, one that for the SGP needed OFF to get a DCOP
without gmin stepping, and that was translated into what should have
been an equivalent VBIC model.  "Fixing" the base-collector voltages
to zero during initFix was a bad thing, as it caused the device to
apply voltage limiting at later steps --- thereby becoming
"unconverged" and turning initFix back on.  Initializing *only* the
base-emitter voltages worked very nicely, but smelled bad.

In talking to @tmei, it seems that what was *really* asked for in
Xyce/Xyce#354 was that the vbic not initialize voltages AT ALL when
OFF was provided.

This commit does just that.  It simply adds an OFF parameter to VBIC,
and modifies the junction voltage initialization code (which is only
applied when initJctFlag_ is true, which is only at the first Newton step of
the DCOP) so that it is skipped entirely if OFF is given.  Thus, the
DCOP is computed directly from the initial  guess or initial condition
given in the netlist, and the junction voltages are not tweaked on the
first step.

This also works well on my simple test case, and may be exactly what
is needed for the circuits that spawned Xyce/Xyce#354.