Unbreak user_plugin demo directory
The "user_plugin" directory is a cheesy demo of the shared-library
plugin feature that Dave Baur created at almost the same time that he
created the plugin capability.

It is a copy of the rlc.va verilog-a file from the
utils/ADMS/examples/toys directory.  The only thing this directory
provides is a Makefile.am to allow the plugin to be built either as an
afterthought (by cding to its build directory and running "make") or
as part of the full Xyce build (by adding "--enable-user-plugin" to
the configure line).

Since it is a real pain in the butt to build plugins this way,
especially for the average consumer of the Xyce/ADMS product, I
eventually wrote the "buildxyceplugin" script to separate plugin
creation from the Xyce build process allowing it to be relatively easy
for a user to make verilog-a plugins (so long as they are feeding it
good verilog-a).

Unfortunately, when I refactored Xyce/ADMS to do away with Sacado, I
changed how one invokes ADMS to get the job done --- all the XML files
have new names, and there are a few extra new XML files needed.  When
I did this, I never bothered to fix the user_plugin directory, because
the user_plugin directory really never served any purpose other than
as a demo, and nobody uses it anymore.

Until today, when a user reported a Xyce build failure that was
entirely do to their having added "--enable-user-plugin" to their
configure line, despite absolutely NO documentation of the plugin
feature recommending this (they probably just pulled it out of the
"--help" text from configure, which documents every possible option in
one line each).  It broke because back in Xyce 7.2 I removed the old
XML files that this plugin directory referenced.

This commit resores user_plugin to the level of usefulness it had
before Xyce 7.2.  Which is to say, it builds, but it's still useless.