Fixes NOX transient nonlinear solver infrastructure
After the transition from NOX to our internal Newton solver as the
default for transient simulation, changes have caused a divergence
in the behavior of the two solvers.  If NOX is used as the transient
solver, several tests will fail due to convergence errors or seg
faults.  While integrating device masking into the nonlinear solvers
many, but not all, of the issues causing this divergence were found.
Included in this commit are the following:

1)  Integration of device masking into the NOX nonlinear solver.
2)  Correction of the behavior when the command line flag for NOX is
    used.  Providing the command line flag "-nox on" would not use
    NOX for transient simulation because the nonlinear manager did not
    process that line correctly.
3)  Correction of the tolerances for nonlinear "near convergence",
    which were swapped in the argument list for the XyceTest object.
4)  Addition of the 2-norm "too small" convergence check which is in
    the DampedNewton nonlinear solver, but not in the NOX convergence
    checks for transient simulation.
5)  Correction of the handling of failures when devices are not
    converged to be consistent with DampedNewton.
6)  Correction of the handling of convergence for 2-level Newton
    solves to be consistent with Damped Newton.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/329