Fixes ROL interface to use new linear algebra abstractions
The current interface to ROL is compatible with Trilinos 12.4.
Yes, that's right, Trilinos 12.4.  It will be updated in the near
future to Trilinos development, to support a desing optimization
project under DARPA.  But first the deprecated ROL vector interface
is being updated to the current linear algebra interfaces in Xyce.
This is merely for the purposes of providing a working example
to the ROL developers.