Moves abstract Linear::Problem object to Linear::Solver base class
Each one of the solver implementations stores the reference to the
abstract Linear::Problem class that is passed in during construction.
Access to the linear problem from any of the solver objects may be
useful during analysis.  So, this repeated pattern has been moved
to the Linear::Solver base class along with an accessor for the
Linear::Problem object.

This avoids the need for indirect manipulation of the linear solver
objects from the nonlinear solver, which is done in the Xyce-ROL
implementation.