Fix Xyce/ADMS ternary op generation
Xyce/ADMS was unnecessarily continuing to emit  templated function
definitions of "adms_ternary_op" and calls to that when using the
ternary operator in analog function context.

This is a relic of our old Sacado usage, and is no longer necessary. I
had already removed all use of this "adms_ternary_op" in all contexts
other than analog functions, but for some reason never finished
removing it in analog function context.

Furthermore, the templating was never perfectly general (it assumed
either that both the "iftrue" and "iffalse" expressions were of the
same non-POD type (e.g. a Sacado FAD type) or that one was a double
and one was something else with a well-defined conversion for
references.  Sometimes a user would find a combination of expressions
that was not compatible with this incomplete template structure.  That
happened over the holiday break (the subject of github issue 49).

The correct fix for this is to remove usage of "adms_ternary_op" in
analog function context and simply emit the C++ ternary operator that
corresponds to the Verilog-A ternary operator (that is, the same
output as the input).

Xyce#358
closes GH-49