Regenerate models after Xyce/ADMS change
Xyce/ADMS was updated, and so all the models generated from Verilog-A
using ADMS need to be regenerated.

The primary impact of the backend change on Xyce is to remove a bunch
of unused templated function definitions, but in one case (BSIM-SOI
4.6) a single line of analog function code is changed from calling
"adms_ternary_op" to simply using the C++ ternary operator.

Xyce#358
GH-49