Add "dt" as a special variable in expression lib
The expression library already tracks "dt" (the time step) as a
"special" AST node in expression trees.  It does this as part of its
support for the ddt and sdt operators.

For debugging purposes it can be convenient to access "dt" directly in
an expression.  This commit adds that feature, by adding a new token
"TOK_DT" and a new rule in the parser to handle it.