Fix logQueuedData and outputFailedStepData for parallel execution
Ting reported (some time ago) that the data output from failed steps,
particularly the node name associated with the inf norm, can be incorrect
in parallel.  Instead, the node name will be printed out as "N/A".
This is not useful for developers that are trying to debug problematic
circuits in parallel.

The issue is that the max/inf norm returns the GID of the node that has
the largest value.  Then the code tried to use that GID to obtain the
character string for that node from topology, which only maps local
node names to LIDs.  The fix is simply to convert the GID to an LID,
where only one processor will have a valid LID (the rest will have -1).
The processor with the valid LID then broadcasts the character string
name of the node to the other processors.

This came up as part of the nonlinear solver work:

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/329