Adds capability to query the linear problem for NaNs
For some challenging ASICs, the nonlinear solver and linear solver
observe NaNs in the Jacobian and/or residual vector.  A relatively
inexpensive check has been inserted, first in Amesos, to determine
if a NaN was encountered in the linear solve.  While it would be
nice that the linear solvers fail upon observing a NaN, they most
often don't.  It's best to trace the NaNs to the Jacobian and residual
vector because then it is possible to output the matrix/vector GIDs.
This facilitates easier debugging for developers.

Found as part of this bug:

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/329