Modifies convergence check to address NaNs in Jacobian/residual vector
The linear solver can now fail due to a detection of NaNs in the
Jacobian matrix and/or residual vector.  If it does, the nonlinear
failure will be a NaN failure, not a linear solver failure.  The
GIDs for the matrix/vector entries can be retrieved from the solver
and printed out when verbosity is turned on for both the internal
Newton solver and NOX solver.

This is done in part for issue:

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/329