Change location of verilog attributes
It was pointed out to us today that Xyce/ADMS accepts Verilog-A
"attributes" that appear *after* the thing being attributed as well as
when they appear as prefix.  That's because ADMS has always accepted
this method.

The Verilog-A standard states that they are supposed to be prefixes,
and ADMS, since it predates that standard, is accepting more than the
standard allows.

Many of our older legacy models (including many that were CMC devices
that predate the use of CMC standard macros) use attributes in the
non-standard way, and *ALL* of our demo "toys" do.

I am changing all of our demo "toys" to use the standard way, and the
Xyce/ADMS users guide has been updated to say that while ADMS accepts
suffixed attributes, it is non standard and new models should be
written with attributes as prefixes if they are intended to be
portable to other Verilog-A compilers.

The handling of attributes in this manner is done at the deeper levels
of ADMS itself, not in Xyce/ADMS.  ADMS will not be "fixed" to stop
accepting prefixed attributes (or at least, *I* will not fix ADMS to
do it), and this is NOT a mistake in Xyce/ADMS, so there's nothing to
fix there.  But we don't have to encourage users to violate the
standard, even though we accept those violations.