Fixes unchecked access fix to logQueuedData()
An unchecked access of the node name vector from topology to obtain
the node of largest absolute value can cause seg faults if (for some
reason) that index was not valid.  The index is checked for its
validity now in the logQueuedData method.

This corrects the earlier commit to address issue:

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/329