Add new device options to support issue 14 SRN
This commit adds new options that are specified on the
".options device" line, and they apply to a new
photocurrent model formulation.

Most importantly, there is now an option that allows one to
choose between different rad model formulations.

Other new settings include:
 - tolerance parameters used by the new formulation
 - a boolean enabling variable masking in the old formulation

This commit is in support of an internal issue:
Xyce/Xyce_SandiaModels#14

I initially thought about implementing these options as device instance
or model-level parameters.  But I want to be able to set them
for entire circuits with a single netlist command, and so
this seemed a better place to put them.