Make OFF parameter both instance and model
With this commit, a VBIC model card may have "OFF=1" and all devices
that use that model card will skip junction voltage initialization
altogether.

Philosophically, I do not like this sort of thing, as "OFF" in the BJT
is meant to be used to improve convergence by saying "make this one
specific BJT be started in an OFF state", whereas putting it in the
model card makes all device of the same type start in that state no
matter how they're connected or what environment they're in.

But given how many other simulators completely blur the distinction
between model and instance parameters (or make no distinction at all),
this is just an old man yelling at a cloud.

Xyce/Xyce#354
Xyce/Xyce!32