Add timeint method enum, available in device pkg
This commit is motivated by my needing to know the current integration
method in the device package.  The device pacakge already had access to
a lot of time integration details, but lacked this one.  This is useful
for an internal issue I am working on.

As part of this, I also created an enum to represent the different
methods.  I did this to:

1. remove hardwired numbers in lines.  So now lines like this:

  initialIntegrationMethod_(7),  //constructor
  integrationMethod = 7; // elsewhere ...

are replaced by:

  initialIntegrationMethod_(TimeIntg::methodsEnum::ONESTEP),

  integrationMethod = Xyce::TimeIntg::methodsEnum::ONESTEP;

2. to get rid of a lot of unncessary #include statements,
such as #include <N_TIA_NoTimeIntegration.h>, which was in many,
many files.  But, it was only there to support lines like:

  baseIntegrationMethod_ = TimeIntg::NoTimeIntegration::type;

which are now replaced by:

  baseIntegrationMethod_ = TimeIntg::methodsEnum::NO_TIME_INTEGRATION;

Note, ONESTEP is still equals 7, and GEAR is still 8, but they are
easier to understand and are easier to grep.

For historical reasons, time integration methods are indexed
to the following integers:

0 - no time integration
7 - one step  (trap)
8 - gear

All the other numbers from 1-6 refer to methods that don't exeist
anymore.

Usually, when people set the time integration method from the netlist,
it is set with a string, such as method=trap or method=gear.  But, there
are still some netlists in the test suite that specify them via an
integer.  So, one sometimes encounters "method=7" or "method=8".
This commit doesn't change that, so those tests still pass.

But, if we ever want to change Xyce so that method=int doesn't work
anymore (and users must use method=string) , or to use *different*
integers for method=trap and method=gear, it will now be a little
easier to do.