Improve robustness of source stepping algorithm
Issue #355
"Make it possible to turn off initial junction voltages globally in
device package w/o turning off voltlim"

Issue #363
"Extend source stepping in Xyce to apply scalar to ISRC devices as well
as Vsrc devices"

This commit came out of work to improve source stepping.  I
was attepting to run a customer circuit that was having convergence
problems at the DCOP.  It seemed to run OK if I skipped the DCOP
and ramped all the sources in transient, which implied that source
stepping (via LOCA) should also work.

However, it did not run with source stepping via LOCA, because it
would fail the very first step.  In the homotopy method, if the solver
fails the first step, it has no recourse.  It cannot cut the stepsize
and try again, because it hasn't yet taken any kind of step.  So the
algorithm just fails.

For similar reasons, the DCOP Newton solve is less robust than transient
Newton solves.  For transient solves the solver can adjust the time step
and try again, but for the DCOP it cannot.

My hope was it would converge if I fixed issue #355, which was to ensure
devices were all in the OFF state when running source stepping.  That
way, the initial norms of the first souce step should be perfectly zero.
In that case, the initial step should be trivial to solve, and the
second LOCA step is the first one involving real work by the solver.
And, if that second LOCA step has trouble, unlike the first step,
it can adjust by cutting the stepsize.

But, even with everything set to "off", it still didn't work, and the
initial norms were NOT zero.  As it turns out, there was still a
nonzero source present.  But, it was an Isrc rather than a Vsrc.

It turns out that our "source-stepping" algorithm, as implemented in Xyce,
only affected Vsrcs and not Isrcs.  But to be robust for my circuit it needed
to scale both kinds of source.  Once I modified the code to scale Isrcs as
well as Vsrcs, I was able to get the behavior I wanted.  ie, the initial
LOCA source step showed this nonlinear solver behavior:

||F|| = 0.000e+00  step = 0.000e+00  dx = 0.000e+00
||F|| = 0.000e+00  step = 1.000e+00  dx = 0.000e+00 (Converged!)

I checked the source-stepping algorithm in spice3 (which ours was
based on) and it scales both Vsrcs and Isrcs.  So, this change is
consistent with our original intent.

Note, the particular circuit I was running was pretty fragile, and it
needed BOTH fixes (for 355 and 363) to successfully solve.  If either
fix was disabled, it failed to converge.  It could not tolerate *any*
source in the system on the first source step.

Here are the code details of this commit:

For turning all the devices off, I have added an boolean variable,
"disableInitJctFlag" to the device options class.  If this variable is
set to "true" then the solver_state.initJctFlag_ boolean can never be
true, and the initial junction voltages can never be applied.

I have also added an accessor function to the device package, which
allows this variable "disableInitJctFlag" to be set and unset.  That
funciton is called from the "sourceSteppingSolve" function in the
Xyce::Nonlinear::N_NLS_NOX::Interface class.  That way it is now part of
Xyce's standard DCOP strategy, and also used whenever the user manually
specifies the built-in source stepping algorithm.

I have also added a way to set this manually from the netlist.  There is
a new device options command:

.options device all_off=true

which can be used to set the "disableInitJctFlag" to true or false.  So,
if one wants to do a manual version of source stepping or
pseudo-transient, this can be used.

For the other change, which applies source scaling to the ISRC device,
that only required me to modify the "VsrcScaleParam" class in the
N_DEV_ArtificialParameters.C file.  The name of that class is now
somewhat incorrect, since it doesn't just apply to Vsrcs.  The name is
connected to a hardwired artificial parameter name VSRCSCALE, which now
is a somewhat inaccurate name.  I did not bother changing the name of
the parameter or class, however.  They remain in place for now.

I don't think most users ever directly invoke the VSRCSCALE parameter
from the netlist.  But if I were to change the name of the parameter,
then any netlist that did so would break.  So that is why I have not
changed the name.