Fixes use of wrong output ID for parallel output of transient failures
In commit eff3dd2e the logic in the output of transient failures was
corrected for parallel execution so that it would print the correct
node name, regardless of the processor it was on.  Unfortunately,
while the logic to obtain that output ID was correct, it was not being
correctly used to get the node name on the processor that owned the
GID.  This was causing seg faults in parallel when verbosity was
enabled, but is fixed now.

Fixed as part of issue:

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/329