Small mod to allow .PRINT homotopy with .AC
The parser has a check in it to ensure that the .PRINT type matches the
analysis type.  This check is a bit klunky, and periodically I attempt
to use a .PRINT line that it (incorrectly) doesn't like.

Today I was running an .AC circuit and I was using various homotopy
methods to get thru the DCOP calculation, which is always performed
before the linearized AC analysis.  So, I added a .PRINT HOMOTOPY line.
Xyce didn't like this and exited with an error.

This commit modifies the code that performs that error check so that
this incorrect error will no longer happen.