Make ADMS generate correct code for IE in mosfets
Some mosfets have a fourth node named "E".  ADMS did not know what to
do with this and would emit code only so that I4() could be used to
print the lead current.

This breaks wildcards such as IE(*), because that wildcard won't find
any E nodes for mosfets generated by ADMS, but will find them for
BSIM-SOI 3 (the only non-ADMS MOSFET that has an E node).

BSIM-CMG and BSIM-SOI 4 both have E nodes for the fourth node.  This
commit changes ADMS so that any MOSFET with an E node (that is NAMED E
in the module declaration) will have IE() work.

The only devices in Xyce impacted by this change are BSIM-CMG and
BSIM-SOI 4 (4.5.0 and 4.6.1).

Xyce#365