Tweak description of MOSFET print wildcards
The output of IE() wildcards claimed that it was supported for SOI
devices, but that was not true for BSIM-SOI 4.x devices until today.

Now it is true, and we say so.  It is also legal for BSIM-CMG now.

IB() is only supported for BSIM-SOI 3, and will not be added for
BSIM-SOI 4.
Xyce#365