allow .STEP to work w/integer params
Xyce/Xyce_SandiaModels#16
Xyce/Xyce_SandiaModels#17

This commit is for a couple of internal issues, having to do with
certain models not working sufficiently with parameter-modifying
analyses such as .STEP.

Prior to this commit, analyses like .STEP could only be applied to
real-valued parameters, like a resistance.  If the user wanted to sweep
over an integer-valued parameter, they weren't allowed to.

Most integer parameters probably should NOT be changeable in this way.
For example, right now the LEVEL number of a model should not be
changeable.   The code infastructure simply cannot support it.

However, there are the occasional parameter, which are integers, which
are valid inputs that should be sweepable.   This commit allows integers
to be swept, but not unconditionally.  For an integer parameter to be
allowed in .STEP, that parameter must have been declared as "mutable" in
the parameter description class.

So far, I've only applied this mutable description to one integer
parameter in the Sandia Models directory.  So, for most purposes,
this commit won't change code behavior.