Make "B" lead currents for mosfets printable
Until now, ADMS-generated MOSFETS that have "B" nodes declared in
their module declarations could not have lead currents for the B node
printed using "IB()".  They would be printable only by using the In()
notation, where "n" is their position on the instance line.

This commit fixes that.  ADMS-generated MOSFETs now support ID, IG,
IS, IB, and IE just as the standard legacy models do.

If a node is optional, its lead current is only available if the node
is used on the instance line.  This is primarily only the case for the
body nodes of SOI MOSFETs.

Xyce#366