Extend NaN/Inf index probing to residual vector in nonlinear solver
It is useful, for debugging purposes, to identify which variable has
triggered the NaN/Inf failure in the nonlinear solver.  Initially,
this identification was only performed in the linear solver, which
would catch NaNs/Infs in the linear system.  Now, it is being extended
to the residual vector from the nonlinear solver.  This probing is only
performed if the nonlinear solver verbosity is enabled and it will
output the GID of any variables in the residual vector that are NaN/Inf.

This work supports https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/329