Fix units in DIODE_CMC reference guide tables
The DIODE_CMC authors used "unit=" attributes to identify units of
model and instance parameters.  This is unrecognized by Xyce/ADMS and
is not what other standard models use ("units=").

DIODE_CMC is basically the same as JUNCAP2000, which doesn't have the
same problem.

This leads to units being missing in the tables of parameters in the
reference guide.  This commit makes a straight text substitution of
"unit=" with "units=" where it matters, and regenerates the tables.

I have also added a "xyceDeviceName" attribute to the module so the
table identifies the device as "DIODE_CMC 2.0.0" instead of "ADMS
diode_cmc" now.

There was also a mistake in the text in the reference guide where
reference was made to the "level 200" diode when "level 2002" was
intended.

README_Xyce has been updated to reflect the fact that the model as
saved in our repo is not the same as the one we received, documenting
this text substitution.

Xyce#368