Fix for out-of-bounds bug in expr library
Issue #369
"Out-of-bounds data storage in "dx2" capability of expression library"

The dx2 function (or set of functions, one for each ast node), was
implemented about 8 months ago to make the expression library more
efficient.  The originaly dx function was used by the expression library
to produce derivatives, but it only produced a single derivative per AST
traversal.

The dx2 functions, in contrast, compute all the reqested derivatives in
a single AST traversal, and this which is more efficient.

Xyce's use of derivatives is limited to device models such as the Bsrc,
and many of those uses only have 1 or 2 derivatives.  Also, in practice,
many syntax trees are small.  So, this wasn't a huge speed optimization,
but it was the right thing to do.

Anyway, in order to do these traversals for multiple derivatives, an
array of derivatives has to get passed around.  And, for AST nodes that
have mutliple arguments, it is necessary to create and use temporary
derivative arrays during the traversal.

I had set up these temporary derivative arrays to only be allocated one
time, as I imagined that once an expression was set up, the number of
derivatives was constant.  However, this is not the case when using
.FUNCs.  In that case, a .FUNC can be called with different arguments,
which may have different numbers of required derivatives.

I was given a test case yesterday that did exactly this.  The result was
a core dump on RHEL7.

The fix was simply to change the logic used with temporary derivative
arrays.  Now, in the AST nodes that need them, these arrays are local to
the dx2 function, and thus are allocated every time they are used.