Adds accessor for the residual vector and Jacobian from the nonlinear solver
There was an ifdef in the nonlinear solver that would allow ROL to access
the residual and Jacobian from the nonlinear solver.  This is allowable and
has been added to the nonlinear solver interface in a standardized way that
is no longer ifdef'd.

Changes to the DC Optimization classes for objective and constraints has been
updated to reflect this change.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/370