Fix for AC .SENS efficiency
Issue #364
"AC sensitivities for nonlinear parameters are very slow"

This commit (mostly) fixes this issue.  It do so by completely
re-arranging the computations required to get the right-hand-side of the
sensitivity calculation.

For AC sensitivities, you have to solve an equation like this:

J * x' = b' - J' * x

where x' = dx/dp, where p is a parameter and x is the AC
solution vector.  Most of the work in this commit has to
do with the matrix derivative J' = dJ/dp.

The vast majority of the work needed to produce J' can be done a single
time.  Previously, I was doing it over and over again.

I had assumed that most of the expense was in computing the DC
sensitivities , which is part of the process.  But it turned out that
most of the expense was in matrix arithmetic used to create a block dJdp
matrix from the DC dCdp and dGdp matrices.

Fortunately, it was possible to amortize most of this work and do it a
single time.  Now, when sweeping over frequency, dJdp is only updated by
unscaling and re-scaling the C-blocks of dJdp by omega, where:

omega = 2 * pi * freq_

This scaling procedure is very cheap, compared with what Xyce was doing
before.