Fix for AC sensitivities bug
Issue #372
"AC sensitivities are incorrect when using .param for sensitivity
parameters"

This is a one-line bugfix for an issue I discovered yesterday.  For some
sensitivity analyses involving .param, it is currently necessary to
compute numerical dFdp, dQdp and dBdp finite difference sensitivities.

The dBdp vector was incorrect, because in one spot in the AC class I
needed to zero out the DC B-vector, but I did not.  So, it kept summing
into it.   So, in a number of cases where dBdp should have been zero, it
was big.

This didn't affect normal (non-SENS) AC analysis, as that analysis
doesn't use the DC RHS vectors at all.  They were only relevant to some
use cases of sensitivity analysis.

Fortunately this was easy to fix.