Code re-arrangement
This is inspired by my work on issue #364, to improve the efficiency of
AC sensitivities.

In my previous commit I amortized most of the numerical calculation
of J', so it is only done 1x, rather than many times.

In this commit, I am doing the same for the (rare) cases where I can get
J' analytically.  I initially thought this wasn't worth bothering with,
but I don't feel that way now, given that the block matrix arithmetic
has turned out to be expensive.

This commit also gets rid of some data structures that were not being
used anymore, to save memory.