Fix a couple  of URLs
Our web site migrated to a different CMS, and all the deep links we
used to reference in documentation and/or emails to the google group
are wrong now.

This commit fixes pointers to the documentation page of the web site
and specifically two pointers to the building guide.

We might have to spend time hunting down all the places where we've
provided deep links into the external web site, becuase they'll all be
wrong now.