Fix for AC .SENS in parallel
My recent work for issue #364 broke AC sensitivities in parallel.

In a previous commit, I partially fixed the problem, but not completely.

This commit fixes the rest of it.  The main thing that had to change was
a few places where calls to the matrix function "add" had to be replaced
with calls to "addOverlap" followed by a fill complete.  These calls
were being used to compute the dGdp and dCdp finite difference
Jacobians (on the DC matrices).