Additional refactoring of ROL DC optimization capability
This commit contains the following modifications:

1)  Adds ROL analysis enumeration.
2)  Modifies the extractDCData method by separating the internals
of the parsed line processing into extractDCDataInternals.  This will
provide access to common .DC line processing to higher-level analysis
objects, like ROL.  This facilitates code reuse and reduces the need
to synchronize several copies of the same parsed line processing.
3)  Modifies the ROL analysis lines, creating .ROL_DC for the declaration
of DC analysis to be used by ROL.  The .ROL line has been transitioned
to providing parameters that should be explicitly used by ROL.  In the
future, this can be extended to passing in an XML parameter list, etc.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/370