Fix AC .sens for .param with AC sources
Issue #372
"AC sensitivities are incorrect when using .param for sensitivity
parameters"

I thought I had this fixed, but then I created a test case.  And, what
do you know, it wasn't quite fixed.  I hadn't considered the case
of .param applied to AC sources (the B-vector).

Currently for all types of .SENS, if the sensitivity parameter is a
.param or a .global param, then the only way to get the df/dp, dq/dp and
db/dp from the device models is to do a vector-wide finite difference.

For other kinds of parameters, (like a .MODEL param or instance param)
it is possible to get an analytical derivative, or to do a finite
difference directly in the device.  But for .param and .global param, we
have to do this brute force finite difference.

This won't be the case forever, but it is the case for now.

Anyway, in the logic of the AC sensitivity calculation, the last thing
it attempts (after attempting to get the necessary derivatives
analytically, etc) is to get the RHS numerically. The AC .SENS RHS
(for direct sensitivities) is B' - J' * X.  But, it was only
computing the numerical J', and ignoring the numerical B'.

This commit fixes it so that it also computes the numerical B'.