Separates the DC sweep analysis from the ROL analysis class
This commit performs the following modifications to the ROL analysis:

1.  Creates a new lightweight class, ROL_DC, that is inherited from
the DCSweep class.  This new class uses most of the DCSweep class
implementation and overrides doProcessSuccessfulStep to store the
current solution vector from that step.
2.  Removes all of the copied DCSweep code in the ROL analysis class
and employs the ROL_DC class to perform the DCSweep.  The new ROL_DC
class is passed into the EqualityConstraint_ROL_DC object instead of
the ROL object.

This has the effect of removing most of the ROL analysis code and
clarifying the algorithm flow and interactions between Xyce and ROL.
Additional efficiencies can be achieved by accumulating more information
from Xyce during the doProcessSuccessfulStep method in ROL_DC.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/370